refactor(core): split character module into storage/races/classes/stats
Extract character persistence, races, and stats generation into
dedicated modules while keeping core/character.py as a compatibility
facade for existing imports.
